Moving logic from presenter
