+++ tests/PKSim.Tests/Infrastructure/PopulationPKAnalysesTaskSpecs.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using FakeItEasy;
 using OSPSuite.BDDHelper;
 using OSPSuite.BDDHelper.Extensions;
@@ -6,9 +7,13 @@ using OSPSuite.Core.Domain;
 using OSPSuite.Core.Domain.Data;
 using OSPSuite.Core.Domain.PKAnalyses;
 using OSPSuite.Core.Domain.Services;
+using OSPSuite.Presentation.Services;
+using OSPSuite.Utility.Extensions;
 using PKSim.Core.Mappers;
 using PKSim.Core.Model;
+using PKSim.Core.Model.PopulationAnalyses;
 using PKSim.Core.Repositories;
+using PKSim.Core.Services;
 using ILazyLoadTask = PKSim.Core.Services.ILazyLoadTask;
 using IPKAnalysesTask = PKSim.Core.Services.IPKAnalysesTask;
 using IPKCalculationOptionsFactory = PKSim.Core.Services.IPKCalculationOptionsFactory;
@@ -34,8 +39,13 @@ namespace PKSim.Infrastructure
       protected OutputSelections _outputSelections;
       private PKParameter _pkParameter1;
       private PKParameter _pkParameter2;
-      private IPKValuesToPKAnalysisMapper _pkMapper;
-      private IDimensionRepository _dimensionRepository;
+      protected IPKValuesToPKAnalysisMapper _pkMapper;
+      protected IDimensionRepository _dimensionRepository;
+      protected IStatisticalDataCalculator _statisticalDataCalculator;
+      protected IRepresentationInfoRepository _representationInfoRepository;
+      protected const string _percentileId = "Percentile";
+      protected PercentileStatisticalAggregation _percentileStatisticalAggregation;
+      protected PopulationStatisticalAnalysis _populationStatisticalAnalysis;
 
       protected override void Context()
       {
@@ -46,7 +56,13 @@ namespace PKSim.Infrastructure
          _entityPathResolver = A.Fake<IEntityPathResolver>();
          _pkMapper= A.Fake<IPKValuesToPKAnalysisMapper>();
          _dimensionRepository= A.Fake<IDimensionRepository>();
-         sut = new PKAnalysesTask(_lazyLoadTask, _pkValuesCalculator, _pkParameterRepository, _pkCalculationOptionsFactory, _entityPathResolver,_pkMapper, _dimensionRepository);
+         _statisticalDataCalculator = new StatisticalDataCalculator();
+         _representationInfoRepository = A.Fake<IRepresentationInfoRepository>();
+         _percentileStatisticalAggregation = new PercentileStatisticalAggregation { Selected = true, Percentile = 50 };
+         _populationStatisticalAnalysis = new PopulationStatisticalAnalysis();
+         _populationStatisticalAnalysis.AddStatistic(_percentileStatisticalAggregation);
+         A.CallTo(() => _representationInfoRepository.DisplayNameFor(_percentileStatisticalAggregation)).Returns(_percentileId);
+         sut = new PKAnalysesTask(_lazyLoadTask, _pkValuesCalculator, _pkParameterRepository, _pkCalculationOptionsFactory, _entityPathResolver,_pkMapper, _dimensionRepository, _statisticalDataCalculator, _representationInfoRepository);
 
          _populationSimulation = A.Fake<PopulationSimulation>();
          _outputSelections = new OutputSelections();
@@ -161,6 +177,66 @@ namespace PKSim.Infrastructure
       public void should_have_created_one_pk_analyses_for_each_output_defined_in_the_simulation()
       {
          _results.IsNull().ShouldBeTrue();
+      }
+   }
+
+   public class When_aggregating_pk_parameters_from_individuals : concern_for_PopulationPKAnalysesTask
+   {
+      private IEnumerable<PopulationPKAnalysis> _pkAnalyses;
+
+      class TestQuantityPKParameter : QuantityPKParameter
+      {
+         public float[] _values { get; set; }
+
+         public override float[] ValuesAsArray { get => _values; }
+      }
+
+      protected override void Context()
+      {
+         base.Context();
+
+         var model = A.Fake<IModel>();
+         A.CallTo(() => model.MoleculeNameFor("Organism|PeripheralVenousBlood|Esomeprazole|Plasma (Peripheral Venous Blood)")).Returns("Esomeprazole");
+         A.CallTo(() => model.MoleculeNameFor("Organism|PeripheralVenousBlood|Esomeprazole-2|Plasma (Peripheral Venous Blood)")).Returns("Esomeprazole-2");
+         A.CallTo(() => _populationSimulation.Model).Returns(model);
+         var esomeprazole = A.Fake<Compound>().WithName("Esomeprazole");
+         A.CallTo(() => esomeprazole.MolWeight).Returns(200);
+         A.CallTo(() => _populationSimulation.Compounds).Returns(new[] { esomeprazole, A.Fake<Compound>().WithName("Esomeprazole-2")});
+
+         _populationSimulation.AddCompoundPK(new CompoundPK() { CompoundName = "Esomeprazol-2" });
+         var analysis = new TimeProfileAnalysisChart();
+         analysis.PopulationAnalysis = _populationStatisticalAnalysis;
+      }
+
+      protected override void Because()
+      {
+         var pkParameters = new[]
+         {
+            new TestQuantityPKParameter() { Name = "Name 1", QuantityPath = "Organism|PeripheralVenousBlood|Esomeprazole|Plasma (Peripheral Venous Blood)",   _values = new[] { 0.000f, 0.050f, 0.025f, 0.075f, 1.000f } },
+            new TestQuantityPKParameter() { Name = "Name 2", QuantityPath = "Organism|PeripheralVenousBlood|Esomeprazole|Plasma (Peripheral Venous Blood)",   _values = new[] { 0.00f,  0.25f,  0.75f,  0.50f,  1.00f  } },
+            new TestQuantityPKParameter() { Name = "Name 3", QuantityPath = "Organism|PeripheralVenousBlood|Esomeprazole|Plasma (Peripheral Venous Blood)",   _values = new[] { 0.0f,   2.5f,   5.0f,   7.5f,   10.0f  } },
+            new TestQuantityPKParameter() { Name = "Name 1", QuantityPath = "Organism|PeripheralVenousBlood|Esomeprazole-2|Plasma (Peripheral Venous Blood)", _values = new[] { 0f,     0f,     0f,     0f,     0f } },
+            new TestQuantityPKParameter() { Name = "Name 2", QuantityPath = "Organism|PeripheralVenousBlood|Esomeprazole-2|Plasma (Peripheral Venous Blood)", _values = new[] { 0f,     0f,     0f,     0f,     0f } },
+            new TestQuantityPKParameter() { Name = "Name 3", QuantityPath = "Organism|PeripheralVenousBlood|Esomeprazole-2|Plasma (Peripheral Venous Blood)", _values = new[] { 0f,     0f,     0f,     0f,     0f } }
+         };
+
+         _pkAnalyses = sut.AggregatePKAnalysis(_populationSimulation, pkParameters, _populationStatisticalAnalysis.SelectedStatistics, "Esomeprazole");
+      }
+
+      [Observation]
+      public void should_aggregate_correctly()
+      {
+         _pkAnalyses.ShouldNotBeEmpty();
+         A.CallTo(() =>_pkMapper.MapFrom(
+            200,
+            A<PKValues>.That.Matches(p => p.Values.ContainsItem(0.05f) && p.Values.ContainsItem(0.5f) && p.Values.ContainsItem(5f)),
+            A<PKParameterMode>.Ignored,
+            "Esomeprazole"
+         )).MustHaveHappened();
+         _pkAnalyses.Count().ShouldBeEqualTo(1);
+         var curveData = _pkAnalyses.First().CurveData;
+         curveData.Caption.ShouldBeEqualTo("Esomeprazole-Percentile");
+         curveData.QuantityPath.ShouldBeEqualTo("Organism|PeripheralVenousBlood|Esomeprazole|Plasma (Peripheral Venous Blood)");
       }
    }
 }
